bug fix
I didnt realise the app would crash when hovering over tiles on the upper x/y bounds, which was cause because it checks for a pillar tile below it (which is out of grid bounds). I've just made it so border tiles dont get highlighted for now. I'll make a proper fix with actual checking later