Zooming out added
Can now zoom in and out with the I and O keys. At the minute, it just checks that the tile_scale value is within the bounds of 2 numbers I chose at random. Would like to add specific zoom levels at some point. (maybe ill do that, and this gradual zoom can still be used for mouse wheel movement)